* commit '761734fd3c4e8699f80bf30e6f34b75ad7f4afe1':
  放款前流标bug修改

+++ src/main/java/com/gqhmt/fss/architect/fuiouFtp/service/BidSettleService.java
@@ -112,7 +112,7 @@ public class BidSettleService {
 
         FundOrderEntity fundOrderEntity = paySuperByFuiou.createOrder(toEntity, loanEntity.getPayAmt(), GlobalConstants.ORDER_SETTLE_NEW, loanEntity.getId(), GlobalConstants.BUSINESS_SETTLE, "2");
 
-        Map<Integer, String> map = fuiouPreauthService.getContractNo(bid.getId().longValue());
+        Map<Long, String> map = fuiouPreauthService.getContractNo(bid.getId().longValue());
         BigDecimal bonusAmount = BigDecimal.ZERO;
 
         List<FuiouFtpColomField> fuiouFtpColomFields = new ArrayList<>();
